Cleaned up code errors and moved into new files
